Move MarkToSelectionMode and SelectionToMarkMode enums to filedata.h
Remove unused typedefs.h includes
